+++ resources/lang/en/flash_message.php
@@ -12,4 +12,14 @@ return [
     'link_list.deleted' => 'Link list was deleted successfully!',
     'link_list.not_deleted' => 'Link list could not be deleted!',
 
+    /**
+     * Announcement
+     */
+    'announcement.created' => 'Announcement was created successfully!',
+    'announcement.not_created' => 'Announcement could not be created!',
+    'announcement.updated' => 'Announcement was updated successfully!',
+    'announcement.not_updated' => 'Announcement could not be updated!',
+    'announcement.deleted' => 'Announcement was deleted successfully!',
+    'announcement.not_deleted' => 'Announcement could not be deleted!',
+
 ];