+++ resources/views/announcement/delete.blade.php
@@ -1,24 +1,19 @@
 @extends('layouts.app')
 
-@section('title', 'Kontakte verwalten')
+@section('title', trans('ui.delete_announcement'))
 
 @section('content')
     <div class="container">
         <div class="row justify-content-center">
             <div class="col-md-12">
                 <div class="card card-default">
-
                     <div class="card-header">
-                        Kontakte verwalten
+                        {{ trans('ui.delete_announcement') }}
                     </div>
                     <div class="card-body">
-                        <p><strong>Kontakte: </strong><br>
-                            <a href="{{ route('contacts.create') }}">Kontakt
-                                hinzufügen</a>
-                        </p>
-
-                        @include('partials.contact.index')
+                        @include('partials.announcement.delete')
                     </div>
+
                 </div>
             </div>
         </div>